refactor: migrate basic code about module into framework
