+++ src/framework/log/interface.py
