+++ src/framework/manager/utils.py
