+++ src/module/interface.py
@@ -1,0 +1,6 @@
+from abc import ABCMeta
+from framework import BasicModule as ModuleInterface
+
+
+class BasicModule(ModuleInterface, metaclass=ABCMeta):
+    pass
